feat: nickname mode refactoring
