+++ src/main/resources/FurryNames.json
@@ -1,0 +1,177 @@
+{
+    "type": "me.ddivad.hawk.dataclasses.FurryNames",
+    "names": [
+        "Frostyclaw",
+        "Luckyeyes",
+        "Copperpelt",
+        "Softfur",
+        "Silentwhisker",
+        "Darkmug",
+        "Rainwhisker",
+        "Cleverheart",
+        "Amberpaw",
+        "Boldshadow",
+        "Mistycoat",
+        "Rapidfeet",
+        "Sneakyclaw",
+        "Darksnout",
+        "Sparklingfur",
+        "Sharpclaw",
+        "Ravencrown",
+        "Luckytail",
+        "Thunderwhisker",
+        "Softyelp",
+        "Snowyclaw",
+        "Swiftfeet",
+        "Crystalpelt",
+        "Fierceheart",
+        "Flamepaw",
+        "Brightclaw",
+        "Leafcoat",
+        "Jollytooth",
+        "Stormyclaw",
+        "Wildyelp",
+        "Emberwhisker",
+        "Humblepelt",
+        "Smokyfur",
+        "Richmug",
+        "Goldenclaw",
+        "Silenttail",
+        "Mistyclaw",
+        "Calmheart",
+        "Midnightfur",
+        "Snowytooth",
+        "Honeywhisker",
+        "Lazytail",
+        "Sablepelt",
+        "Darkclaw",
+        "Braveclaw",
+        "Swiftmug",
+        "Fuzzywhisker",
+        "Braveheart",
+        "Dustyclaw",
+        "Rustytail",
+        "Thundercoat",
+        "Calmfeet",
+        "Snowpaw",
+        "Coldyelp",
+        "Fluffywhisker",
+        "Fierceclaw",
+        "Dawnfur",
+        "Quickheart",
+        "Nightclaw",
+        "Wiseeyes",
+        "Rapidclaw",
+        "Sillytail",
+        "Mistywhisker",
+        "Richshadow",
+        "Gentlecoat",
+        "Swiftmug",
+        "Brightclaw",
+        "Windyheart",
+        "Sneakypelt",
+        "Strongclaw",
+        "Twilightfur",
+        "Wiseyelp",
+        "Majesticpaw",
+        "Darkfeather",
+        "Snowpelt",
+        "Coldtoes",
+        "Windyfur",
+        "Softyelp",
+        "Stormyclaw",
+        "Humbleheart",
+        "Tundrapelt",
+        "Speedytooth",
+        "Breezywhisker",
+        "Rustymug",
+        "Goldenwhisker",
+        "Jollyclaw",
+        "Sandycoat",
+        "Boldyelp",
+        "Lunarclaw",
+        "Richpelt",
+        "Sablewhisker",
+        "Quicktail",
+        "Brightcoat",
+        "Braveheart",
+        "Crimsonpaw",
+        "Wittyheart",
+        "Thunderpelt",
+        "Fierceclaw",
+        "Cloudyclaw",
+        "Silentmug",
+        "Swiftcoat",
+        "Windytooth",
+        "Midnightwhisker",
+        "Sillyclaw",
+        "Smokyfur",
+        "Humblepelt",
+        "Silverclaw",
+        "Darkfeather",
+        "Mistywhisker",
+        "Swifttail",
+        "Duskycoat",
+        "Richyelp",
+        "Crystalpelt",
+        "Speedyclaw",
+        "Fluffywhisker",
+        "Darkheart",
+        "Ravencrown",
+        "Softclaw",
+        "Thunderpaw",
+        "Snowytail",
+        "Dawnclaw",
+        "Rustyheart",
+        "Mistyclaw",
+        "Wildclaw",
+        "Gentlewhisker",
+        "Boldtooth",
+        "Sablepaw",
+        "Softtail",
+        "Majesticfur",
+        "Rustyeyes",
+        "Frostclaw",
+        "Coolpaws",
+        "Moonwhisker",
+        "Twinkletoes",
+        "Sablefur",
+        "Wildheart",
+        "Spirithawk",
+        "Gloomyswoop",
+        "Thunderfox",
+        "Mightyfang",
+        "Mistysnout",
+        "Dreamyhowl",
+        "Goldenpelt",
+        "Richmane",
+        "Fireflyer",
+        "Quickwings",
+        "Swiftbunny",
+        "Furryfeet",
+        "Wildwhisker",
+        "Roamingeyes",
+        "Mudpaw",
+        "Grittyclaw",
+        "Stonefang",
+        "Hardhide",
+        "Cloudchaser",
+        "Fluffytail",
+        "Leafylynx",
+        "Sleekears",
+        "Blazingtiger",
+        "Fiercegrowl",
+        "Junglejaguar",
+        "Sneakypaws",
+        "Silverswift",
+        "Swiftclaw",
+        "Thornystud",
+        "Proudmane",
+        "Smokysnout",
+        "Greyhowl",
+        "Riverscale",
+        "Riverheart",
+        "Stormysky",
+        "Thunderpaws"
+    ]
+}
