+++ src/main/resources/bot.properties
@@ -1,4 +1,4 @@
-#Sat Apr 01 01:05:01 IST 2023
+#Sat Apr 01 18:01:31 IST 2023
 name=Hawk
 description=A bot to add and maintain a symbol as a prefix or suffix in staff names.\n
 version=3.0.0-RC1
